Nice update.
Added try-catch blocks for all commands, and many methods, so now,
stackTraces should be really uncommon.

Added "prefix" to help command, and made it so that if there is no
private channel for the user and the bot, one will be made using a
blocking method instead of queue, because it only resulted in an error 
(trying to send the private message before the channel is set up". 

Added some more text to other help commands: "Use like: <Whatever>"

Signed-off-by: Kakan9898 <glennholsson@gmail.com>