New class DailyDose and new eventListener onGenericEvent in DiscordBot.java.
The class sends a series of messages, max once every 24 hours, with the
top three posts on /r/aww (currently) to all message channels named aww. 

The eventListener is called every time an event is called, and checks if
it was more or exactly 24 hours since the last DailyDose was sent. If
true, DailyDose is called

Signed-off-by: Kakan9898 <glennholsson@gmail.com>